+++ apps/extensions-explorer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # saleor-app-extensions-explorer
 
+## 0.1.3
+
+### Patch Changes
+
+- @saleor/apps-shared@1.16.2
+
 ## 0.1.2
 
 ### Patch Changes