+++ apps/stripe/CHANGELOG.md
@@ -1,5 +1,15 @@
 # saleor-app-payment-stripe
 
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies [9c585ac]
+  - @saleor/apps-otel@2.4.2
+  - @saleor/apps-logger@1.6.6
+  - @saleor/apps-shared@1.16.2
+  - @saleor/apps-trpc@4.0.5
+
 ## 2.8.0
 
 ### Minor Changes